requirements.txt
rasa==3.6.13
rasa-sdk==3.6.2
sqlalchemy==1.4.46
psycopg2-binary==2.9.7
mysql-connector-python==8.1.0
sqlite3
pandas==2.0.3
python-dotenv==1.0.0
pyyaml==6.0.1
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
aiofiles==23.2.1
